fixed typos and unittest
