+++ tests/test_cli_scripts.py
@@ -1,9 +1,12 @@
 import contextlib
-import os
+import os, sys
 import subprocess
 import unittest
 
 from parameterized import parameterized
+
+
+python = sys.executable
 
 
 class TestCliScripts(unittest.TestCase):
@@ -32,7 +35,7 @@ class TestCliScripts(unittest.TestCase):
             # Convert sav to JSON
             run = subprocess.run(
                 [
-                    "python3",
+                    python,
                     "-m",
                     "palworld_save_tools.commands.convert",
                     f"tests/testdata/{dir_name}/{base_name}",
@@ -49,7 +52,7 @@ class TestCliScripts(unittest.TestCase):
             )
             run = subprocess.run(
                 [
-                    "python3",
+                    python,
                     "-m",
                     "palworld_save_tools.commands.convert",
                     f"tests/testdata/{dir_name}/1-{base_name}.json",
@@ -64,7 +67,7 @@ class TestCliScripts(unittest.TestCase):
             )
             run = subprocess.run(
                 [
-                    "python3",
+                    python,
                     "-m",
                     "palworld_save_tools.commands.convert",
                     f"tests/testdata/{dir_name}/2-{base_name}",
@@ -81,7 +84,7 @@ class TestCliScripts(unittest.TestCase):
             )
             run = subprocess.run(
                 [
-                    "python3",
+                    python,
                     "-m",
                     "palworld_save_tools.commands.convert",
                     f"tests/testdata/{dir_name}/3-{base_name}.json",
